requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
drf-yasg>=1.21,<2.0
python-dotenv>=1.0,<2.0
